# [FLINK-14118][runtime]Reduce the unnecessary flushing when there is no data available for flush.