@pnowojski 2019-09-20
Could you provide at least results of `networkThroughput` benchmark with `1000,1ms` parameters before and after the change?

Also could you either show that this change improves some existing benchmarks or to add a new benchmark to cover for this for the future?