@wsry 2019-09-26
> Could you either show that this change improves some existing benchmarks or add a new benchmark to cover for this for the future?

I'd like to add a new benchmark to cover this scenario. Actually, I have added one to verify the problem locally.