@pnowojski 2019-09-26
Thanks for the update @wsry. Could you update this PR once you add the new benchmark?

nitty nit: benchmark results are more readable on github/Jira if you embed them in the code block like this:
```
KeyByBenchmarks.arrayKeyBy1MS thrpt 30 1923.863 ± 78.518 ops/ms
KeyByBenchmarks.tupleKeyBy1MS thrpt 30 3377.401 ± 216.982 ops/ms
``` 
plus add/remove some extra tabs.