@wsry 2019-09-27
@pnowojski Thanks for your tips. It really helps. I will update the pr later when the benchmark is ready. Currently, there is still some problems.