@wsry 2019-10-03
@pnowojski I have updated the PR and I also opened a PR for the flink-benchmark project. https://github.com/dataArtisans/flink-benchmarks/pull/31